requirements.txt
streamlit
google-generativeai
pandas
python-dotenv
scikit-learn
numpy